fix bug of post process after merging merge track_poeple_py, predict_people_py missed track is not handled
DCO 1.1 Signed-off-by: Tatsuya Ishihara <tisihara@jp.ibm.com>
