+++ track_people_py/scripts/track_sort_3d_people.py
@@ -20,10 +20,13 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import signal
+from collections import deque
 import copy
+import signal
 import sys
+import traceback
 
+import numpy as np
 import rclpy
 from rclpy.duration import Duration
 
@@ -42,6 +45,78 @@ class TrackSort3dPeople(AbsTrackPeople):
                                      duration_inactive_to_remove=Duration(seconds=self.duration_inactive_to_remove))
 
         self.buffer = {}
+
+    def postprocess_buf(self, now, alive_track_id_list, center_bird_eye_global_list, color_list):
+        # update queue
+        for (track_id, center3d, color) in zip(alive_track_id_list, center_bird_eye_global_list, color_list):
+            # update tracked people buffer
+            if track_id not in self.track_buf.track_input_queue_dict:
+                self.track_buf.track_input_queue_dict[track_id] = deque(maxlen=self.input_time)
+            self.track_buf.track_input_queue_dict[track_id].append(center3d)
+            self.track_buf.track_color_dict[track_id] = color
+
+            # clear missing time
+            if track_id in self.track_buf.track_id_missing_time_dict:
+                del self.track_buf.track_id_missing_time_dict[track_id]
+
+            # update buffer for observation time
+            if track_id not in self.track_buf.track_time_queue_dict:
+                self.track_buf.track_time_queue_dict[track_id] = deque(maxlen=self.input_time)
+            self.track_buf.track_time_queue_dict[track_id].append(now)
+
+        track_pos_dict = {}
+        track_vel_dict = {}
+        for track_id in alive_track_id_list:
+            if len(self.track_buf.track_input_queue_dict[track_id]) < self.input_time:
+                continue
+            # save position and velocity
+            # use raw position
+            track_pos_dict[track_id] = np.array(self.track_buf.track_input_queue_dict[track_id])[-1, :2]
+            # ues filtered position
+            # track_pos_dict[track_id] = self.tracker.kf_active[track_id]["kf"].x.reshape(1, 4)[0, [0, 2]]
+            track_vel_dict[track_id] = self.tracker.kf_active[track_id]["kf"].x.reshape(1, 4)[0, [1, 3]]
+            if track_id not in self.track_buf.track_vel_hist_dict:
+                self.track_buf.track_vel_hist_dict[track_id] = deque(maxlen=self.input_time)
+            self.track_buf.track_vel_hist_dict[track_id].append((self.track_buf.track_time_queue_dict[track_id][-1], track_vel_dict[track_id]))
+
+        # clean up missed track if necessary
+        missing_track_id_list = set(self.track_buf.track_input_queue_dict.keys()) - set(alive_track_id_list)
+        stop_publish_track_id_list = set()
+        for track_id in missing_track_id_list:
+            # update missing time
+            if track_id not in self.track_buf.track_id_missing_time_dict:
+                self.track_buf.track_id_missing_time_dict[track_id] = now
+
+            # if missing long time, stop publishing in people topic
+            if (now - self.track_buf.track_id_missing_time_dict[track_id]).nanoseconds/1000000000 > self.duration_inactive_to_stop_publish:
+                stop_publish_track_id_list.add(track_id)
+
+            # if missing long time, delete track
+            if (now - self.track_buf.track_id_missing_time_dict[track_id]).nanoseconds/1000000000 > self.duration_inactive_to_remove:
+                del self.track_buf.track_input_queue_dict[track_id]
+                del self.track_buf.track_time_queue_dict[track_id]
+                if track_id in self.track_buf.track_vel_hist_dict:
+                    del self.track_buf.track_vel_hist_dict[track_id]
+                del self.track_buf.track_color_dict[track_id]
+                del self.track_buf.track_id_missing_time_dict[track_id]
+                if track_id in self.track_buf.track_id_stationary_start_time_dict:
+                    del self.track_buf.track_id_stationary_start_time_dict[track_id]
+
+        # add track which is missing, but not deleted yet
+        publish_missing_track_id_list = missing_track_id_list - stop_publish_track_id_list
+        for track_id in publish_missing_track_id_list:
+            if track_id not in self.track_buf.track_input_queue_dict:
+                continue
+
+            last_vel = self.tracker.kf_active[track_id]["kf"].x.reshape(1, 4)[0, [1, 3]]
+
+            # save position and velocity
+            # use raw position
+            track_pos_dict[track_id] = np.array(self.track_buf.track_input_queue_dict[track_id])[-1, :2]
+            # ues filtered position
+            # track_pos_dict[track_id] = self.tracker.kf_active[track_id]["kf"].x.reshape(1, 4)[0, [0, 2]]
+            track_vel_dict[track_id] = last_vel
+        return track_pos_dict, track_vel_dict
 
     def detected_boxes_cb(self, detected_boxes_msg):
         self.htd.tick()
@@ -78,6 +153,7 @@ class TrackSort3dPeople(AbsTrackPeople):
             _, alive_track_id_list, color_list, tracked_duration = self.tracker.track(now, detect_results, center_bird_eye_global_list)
         except Exception as e:
             self.get_logger().error(F"tracking error, {e}")
+            self.get_logger().error(traceback.format_exc())
             return
 
         track_pos_dict, track_vel_dict = self.postprocess_buf(now, alive_track_id_list, center_bird_eye_global_list, color_list)
